Add files via upload
An unpublished paper that I wrote in 2019 that I am releasing into the public domain - a novel idea to counter disinformation, digital addiction and loss of trust. TL;DR 👇🏽

The basic idea - we are evolved for face-to-face communication in which we use *mutual attention* (e.g. through gaze 👁) to regulate / filter communication. This forms a "meta-network" that controls the flow of information.

Digital networks inundate us with information without this meta-signalling and thus overwhelm our evolved cognitive functions. This is the root of the plethora of problems that we are now encountering.

So basically we need to retro-fit the attention economy to make it a "mutual-attention" economy.

In this paper I describe the model of mutual attention, propose a system design for implementing onto existing networks and propose some important use-cases.